refactor: Renamed `FrameCache` to `DataCache`
