+++ src/colorizer/DataCache.ts
